Rename new component and cleanup
