+++ web/src/features/fireBehaviourCalculator/components/FBCInputGrid.tsx
@@ -1,27 +1,20 @@
-import { TextField, Tooltip } from '@material-ui/core'
-import {
-  DataGrid,
-  GridCellParams,
-  GridEditCellValueParams,
-  GridRowId,
-  GridToolbarColumnsButton,
-  GridToolbarContainer,
-  GridToolbarDensitySelector,
-  // GridToolbarExport,
-  GridToolbarFilterButton
-} from '@material-ui/data-grid'
-import InfoIcon from '@material-ui/icons/Info'
-import { Autocomplete } from '@material-ui/lab'
-import { find, isEqual, isNull, isUndefined } from 'lodash'
 import React from 'react'
+import { useTable } from 'react-table'
+import { isNull, isUndefined } from 'lodash'
+import MaUTable from '@material-ui/core/Table'
+import TableBody from '@material-ui/core/TableBody'
+import TableCell from '@material-ui/core/TableCell'
+import TableHead from '@material-ui/core/TableHead'
+import TableRow from '@material-ui/core/TableRow'
 import { FuelTypes } from '../fuelTypes'
-export interface FBCInputGridProps {
+import { buildRowCell } from '../tableRowBuilder'
+
+export interface DatePickerProps {
   testId?: string
-  stationMenuOptions: GridMenuOption[]
-  fuelTypeMenuOptions: GridMenuOption[]
+  stationOptions: GridMenuOption[]
+  fuelTypeOptions: GridMenuOption[]
   rows: FBCInputRow[]
-  updateRow: (rowId: GridRowId, updatedRow: FBCInputRow) => void
-  setSelected: (rowIds: number[]) => void
+  updateRow: (rowId: number, updatedRow: FBCInputRow) => void
 }
 
 export interface GridMenuOption {
@@ -37,55 +30,12 @@ export interface FBCInputRow {
   windSpeed: number | undefined
 }
 
-const buildFBCGridToolbar = () => {
-  return (
-    <GridToolbarContainer>
-      <GridToolbarColumnsButton />
-      <GridToolbarFilterButton />
-      <GridToolbarDensitySelector />
-      {/* <GridToolbarExport /> */}
-    </GridToolbarContainer>
-  )
-}
-export interface DropDownEditProps {
-  label: string
-  options: GridMenuOption[]
-}
-
-interface OptionBoxType {
-  type: 'station' | 'fuelType'
-}
-interface NumberEditProps {
-  value: string
-}
-const GrassCurePercentageEdit = (props: NumberEditProps) => {
-  return (
-    <TextField
-      id="grass-cure-percentage-number"
-      type="number"
-      value={props.value}
-      required={true}
-    />
-  )
-}
-
-const WindSpeedEdit = (props: NumberEditProps) => {
-  return (
-    <TextField
-      id="wind-speed-edit-number"
-      type="number"
-      value={props.value}
-      required={false}
-    />
-  )
-}
-
-const FBCInputGrid = (props: FBCInputGridProps) => {
+const FBCInputGrid = (props: DatePickerProps) => {
   const stationCodeMap = new Map(
-    props.stationMenuOptions.map(station => [station.value, station.label])
+    props.stationOptions.map(station => [station.value, station.label])
   )
 
-  const buildStationOptionFromValue = (value: number) => {
+  const buildStationOption = (value: number) => {
     const label = stationCodeMap.get(value)
     if (isUndefined(label)) {
       return null
@@ -99,7 +49,7 @@ const FBCInputGrid = (props: FBCInputGridProps) => {
 
   const buildFuelTypeMenuOption = (value: string) => {
     const fuelType = FuelTypes.lookup(value)
-    if (isUndefined(fuelType)) {
+    if (isUndefined(fuelType) || isNull(fuelType)) {
       return null
     }
     const option: GridMenuOption = {
@@ -109,138 +59,69 @@ const FBCInputGrid = (props: FBCInputGridProps) => {
     return option
   }
 
-  const optionComboBox = (
-    params: GridCellParams,
-    optionBoxType: OptionBoxType,
-    options: GridMenuOption[]
-  ) => {
-    const { id, api, field } = params
-    const rowToUpdate = find(props.rows, ['id', params.id])
-    if (!rowToUpdate) {
-      return
-    }
+  const data = props.rows.map(row => ({
+    weatherStation: buildStationOption(parseInt(row.weatherStation)),
+    fuelType: buildFuelTypeMenuOption(row.fuelType),
+    grassCure: row.grassCure,
+    windSpeed: row.windSpeed
+  }))
 
-    const currentValue =
-      optionBoxType.type === 'station'
-        ? buildStationOptionFromValue(parseInt(rowToUpdate.weatherStation))
-        : buildFuelTypeMenuOption(rowToUpdate.fuelType)
-
-    // eslint-disable-next-line
-    const handleChange = (_: React.ChangeEvent<{}>, option: GridMenuOption | null) => {
-      if (isNull(option)) {
-        return
+  const columns: any = React.useMemo(
+    () => [
+      {
+        Header: 'Station',
+        accessor: 'weatherStation'
+      },
+      {
+        Header: 'FuelType',
+        accessor: 'fuelType'
+      },
+      {
+        Header: 'Grass Cure %',
+        accessor: 'grassCure'
+      },
+      {
+        Header: 'Wind Speed (km/hr) (Optional)',
+        accessor: 'windSpeed'
       }
-      const editProps = { value: option }
-      api.setEditCellProps({ id, field, props: editProps })
-      api.commitCellChange({ id, field })
-      api.setCellMode(id, field, 'view')
-      const updatedRow = {
-        ...rowToUpdate,
-        ...{
-          [params.field as keyof FBCInputRow]: option?.value
-        }
-      }
-      props.updateRow(params.id, updatedRow)
-    }
+    ],
+    []
+  )
 
-    return (
-      <Autocomplete
-        id={`combo-box-fuel-types-${Math.random()}`}
-        getOptionSelected={(option, value) => isEqual(option, value)}
-        options={options}
-        getOptionLabel={option => option?.label}
-        style={{ width: 300, height: '100%', marginTop: 20 }}
-        renderInput={params => <TextField {...params} variant="outlined" />}
-        onChange={handleChange}
-        value={currentValue}
-      />
-    )
-  }
+  const tableInstance = useTable({ columns, data })
 
-  const updateCellValue = (params: GridEditCellValueParams) => {
-    if (!isEqual(params.field, 'grassCure') && !isEqual(params.field, 'windSpeed')) {
-      return
-    }
-    const rowToUpdate = find(props.rows, ['id', params.id])
-    if (rowToUpdate) {
-      const updatedRow = {
-        ...rowToUpdate,
-        ...{
-          [params.field as keyof FBCInputRow]: params.value
-        }
-      }
-      props.updateRow(params.id, updatedRow)
-    }
-  }
+  const { getTableProps, headerGroups, rows, prepareRow } = tableInstance
 
   return (
-    <div style={{ display: 'flex', height: 300, width: 1200 }}>
-      <div style={{ flexGrow: 1 }}>
-        <DataGrid
-          components={{
-            Toolbar: buildFBCGridToolbar
-          }}
-          checkboxSelection={true}
-          onSelectionModelChange={e => props.setSelected(e.selectionModel as number[])}
-          hideFooter={true}
-          rowHeight={50}
-          columns={[
-            {
-              field: 'weatherStation',
-              headerName: 'Weather Station',
-              flex: 1,
-              type: 'singleSelect',
-              editable: true,
-              renderCell: (params: GridCellParams) =>
-                optionComboBox(params, { type: 'station' }, props.stationMenuOptions),
-              renderEditCell: (params: GridCellParams) =>
-                optionComboBox(params, { type: 'station' }, props.stationMenuOptions)
-            },
-            {
-              field: 'fuelType',
-              flex: 1,
-              headerName: 'Fuel Type',
-              type: 'singleSelect',
-              editable: true,
-              renderCell: (params: GridCellParams) =>
-                optionComboBox(params, { type: 'fuelType' }, props.fuelTypeMenuOptions),
-              renderEditCell: (params: GridCellParams) =>
-                optionComboBox(params, { type: 'fuelType' }, props.fuelTypeMenuOptions)
-            },
-            {
-              field: 'grassCure',
-              headerName: 'Grass Cure %',
-              flex: 0.7,
-              type: 'number',
-              editable: true,
-              renderCell: function numberPicker(params) {
-                return <GrassCurePercentageEdit value={params.value as string} />
-              }
-            },
-            {
-              field: 'windSpeed',
-              // eslint-disable-next-line react/display-name
-              renderHeader: () => (
-                <span>
-                  {'Wind Speed (km/h) (Optional)'}
-                  <Tooltip title="Leave this empty to calculate forecasted/observed wind speed. Add a custom wind speed to influence the calculations">
-                    <InfoIcon aria-label="info"></InfoIcon>
-                  </Tooltip>
-                </span>
-              ),
-              flex: 1.0,
-              type: 'number',
-              editable: true,
-              renderCell: function numberPicker(params) {
-                return <WindSpeedEdit value={params.value as string} />
-              }
-            }
-          ]}
-          rows={props.rows}
-          onCellValueChange={updateCellValue}
-        />
-      </div>
-    </div>
+    <MaUTable {...getTableProps()}>
+      <TableHead>
+        {headerGroups.map(headerGroup => (
+          <TableRow {...headerGroup.getHeaderGroupProps()}>
+            {headerGroup.headers.map(column => (
+              <TableCell {...column.getHeaderProps()}>
+                {column.render('Header')}
+              </TableCell>
+            ))}
+          </TableRow>
+        ))}
+      </TableHead>
+      <TableBody>
+        {rows.map((row, i) => {
+          prepareRow(row)
+          return (
+            <TableRow {...row.getRowProps()}>
+              {row.cells.map(cell => {
+                return (
+                  <TableCell {...cell.getCellProps()}>
+                    {buildRowCell(props, cell, cell.column.id, parseInt(row.id))}
+                  </TableCell>
+                )
+              })}
+            </TableRow>
+          )
+        })}
+      </TableBody>
+    </MaUTable>
   )
 }
